Implement custom msg handler in test to workaround AppVeyor failure
Somehow py34-pyqt5 on AppVeyor test_logging.test_basics fails
with the case where qt logging is disabled and output to stderr is
expected: on AppVeyor no output is generated.

Try to workaround that by explicitly using a custom
message handler in the test that explicitly prints to stderr,
avoiding depending on the default behavior.