Improve logging context handling
Logging context is available with PyQt6. However, given we dropped Qt 4 and the
old qInstallMsgHandler (without a context argument at all), we can now simplify
the code: If a context is available, we use it without any Qt API checks; if
none is available, we don't (and remove the indent from the log message
pattern).

Fixes #364