Docs: suggest direct methods instead of qtbot's
Updates the documentation to use direct methods in most interactions,
as the documentation currently leads users to call `qtbot`'s methods
for normal interactions, such changing the text of a `QLineEdit` or an
item of a `QComboBox`.

Also changed some code examples to use `.. code-block: python`, so they
are formatted by black.

Co-authored-by: Florian Bruhin <me@the-compiler.org>