Add note about PySide6 exception capture and fix tests
PySide6 does its own exception capture during events, re-raising them when control gets back to Python.

This feature causes our own exception capturing mechanism to not work, but it is not needed anyway, because the exception re-raising done by PySide6 contains a better traceback than it is possible for pytest-qt to provide, given it can re-raise at the exact point where control was given back to Python.

Besides noting this in the documentation, also accomodate the tests.